requirements.txt
wheel
boto3==1.9.211
botocore==1.12.211

mysql-connector-python
pandas
requests
Django
djangorestframework
django-environ
markdown
django-filter
django_apscheduler